Module rsaltpack::armor

Reexports

pub use SaltpackMessageType;

Structs

ArmoringStream
Can be used as a streaming interface to armor large amounts of binary data. But it is not recommended to use the armored version to send big amounts of data. It is slow and inefficient.

Range
A (half-open) range which is bounded at both ends.

Constants

NEWLINE_EVERY
The armored string will contain a newline instead of a space every 200 words.

SPACE_EVERY
The armored string will contain a space every 15 characters.

Functions

alphabet
This function converts a zero based digit in base 62 to its ascii equivalent.

armor
This function is a short interface to ArmoringStream. It will convert binary input into the base62 armored version including header and footer.

b32bytes_to_base62_formatted
This function armors one 32 byte block into 43 base62 characters.

Implementations