Function rsaltpack::armor::b32bytes_to_base62_formatted

pub fn b32bytes_to_base62_formatted(raw_in: &[u8], base62out: &mut [u8], space_in: &mut usize, newline_in: &mut usize, full_32: bool) -> usize

This function armors one 32 byte block into 43 base62 characters.

It inserts spaces and newlines, as soon as the counters space_in and newline_in reach zero. It is intended that consecutive calls always get references to the same counters, so the inserted spaces are all equidistant.

If full_32 is true, it will always output 43 characters right aligned, even if the 32 byte input can be represented with 42 characters.

Returns the number of bytes, that have been written into base62out.

Base62out must have place for BUF_SIZE characters, otherwise it will panic.