Struct rsaltpack::armor::ArmoringStream

pub struct ArmoringStream {
    // some fields omitted
}

Can be used as a streaming interface to armor large amounts of binary data. But it is not recommended to use the armored version to send big amounts of data. It is slow and inefficient.

Methods

impl ArmoringStream

fn new(vendorstring: &str, messagetype: SaltpackMessageType) -> Result<ArmoringStream, String>

Create a new streaming interface

fn predict_armored_len(&self, binary_data_len: usize) -> usize

Predicts the total length of armored data including header and footer. Since the armored version is guaranteed to contain only ascii characters, the (armored output as [u8]).len() equals its utf-8 size.

fn armor(&mut self, binary_in: &[u8], last_bytes: bool, armored_out: &mut [u8]) -> Result<(usize, usize), String>

Reads bytes from binary_in and writes the armored version into armored_out If binary_in contains the last bytes that have to be written, set last_bytes to true. Then the footer will be written.

Returns the bytes read from binary_in, the bytes written to armored_out. Call is_done() to figure out if everything (incl. footer) has been written.

fn is_done(&self) -> bool

Returns true if header, data and footer have been written completely.

fn cancel(&mut self)

Stops writing and reading anything. is_done will be true.