Module rsaltpack::compose

Asymmetrically encrypt or sign data for multiple recipients using the saltpack format.

Usage

use rsaltpack::KeyPair;
use rsaltpack::compose::Saltpack;
let recipients = vec![KeyPair::gen().p];

let data = b"If I put a usb stick into my ass, \
                    does it make me a cyborg already?";

// using None as first parameter for anonymous sender
let mut saltpack = Saltpack::encrypt(None, &recipients);

use std::io::Write;
saltpack.write_all(&data[..]).unwrap();
saltpack.flush().unwrap();

use std::io::Read;
let mut encrypted = vec![0u8; 0]; // totally binary data
saltpack.read_to_end(&mut encrypted).unwrap();

assert!(encrypted.len() > 100);

You can use Saltpack as a buffer. Saltpack encrypts in chunks of 1 MB (1000000 bytes). Flush after end of write, otherwise a panic will occur. Also read all data, otherwise a panic will occur.

use std::io::Write;
use std::io::Read;
let mut encrypted = vec![0u8; 0];

saltpack.write_all(&data[0..500_000]).unwrap(); // only half chunk
saltpack.read_to_end(&mut encrypted).unwrap();
assert!(encrypted.len() < 1000); // only header written

saltpack.write_all(&data[0..1_000_000]).unwrap();
saltpack.read_to_end(&mut encrypted).unwrap();
assert!(1_000_000 < encrypted.len() && encrypted.len() < 1_005_000); // 1 chunk written

saltpack.flush().unwrap();
saltpack.read_to_end(&mut encrypted).unwrap();
assert!(1_500_000 < encrypted.len() && encrypted.len() < 1_501_000); // 1.5 chunks written

ASCII armor (only a-z A-Z 0-9 and .)

let mut saltpack = Saltpack::encrypt(None, & vec![KeyPair::gen().p]);
use std::io::Write;
saltpack.write_all(b"secretData").unwrap();
saltpack.flush().unwrap();

let mut armored = saltpack.armor("").unwrap();
println!("{}", armored.to_string()); // read all
// or use std::io::Read on armored to get ascii character as binary string.

Take a look at the Saltpack struct.

Structs

ArmoredSaltpack

Saltpack

Functions

encrypt_to_binary